+++ tests/nugraph.Tests/NugraphAssertions.cs
@@ -27,14 +27,11 @@ public class NugraphAssertions(NugraphResult instance, AssertionChain chain) :
     protected override string Identifier => "result";
 
     /// <summary>
-    /// Asserts that the nugraph result is successful.
-    /// <list type="bullet">
-    /// <item>Ensures that exit code is 0.</item>
-    /// <item>Ensures that stderr is empty.</item>
-    /// <item>Ensures that stdout matches the provided pattern.</item>
-    /// </list>
+    /// Asserts that the nugraph result matches the exit code, stdout and stderr content.
     /// </summary>
-    /// <param name="stdOutPattern">The expected pattern that must be written on stdout.</param>
+    /// <param name="exitCode">The expected exit code.</param>
+    /// <param name="stdOutPattern">The expected pattern that must be written on stdout or an empty string to assert that nothing was written.</param>
+    /// <param name="stdErrPattern">The expected pattern that must be written on stderr or an empty string to assert that nothing was written.</param>
     /// <param name="because">
     /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
     /// </param>
@@ -42,7 +39,7 @@ public class NugraphAssertions(NugraphResult instance, AssertionChain chain) :
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
     [CustomAssertion]
-    public AndConstraint<NugraphAssertions> Succeed(string stdOutPattern, [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
+    public AndConstraint<NugraphAssertions> Match(int exitCode = 0, string stdOutPattern = "", string stdErrPattern = "", [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         string[] failures = [];
 
@@ -51,9 +48,19 @@ public class NugraphAssertions(NugraphResult instance, AssertionChain chain) :
             .ForCondition(result =>
             {
                 using var scope = new AssertionScope();
-                result.ExitCode.Should().Be(0, because, becauseArgs);
-                result.StdErr.Should().BeEmpty(because, becauseArgs);
-                result.StdOut.Should().MatchEquivalentOf(stdOutPattern, opt => opt.IgnoringNewlineStyle(), because, becauseArgs);
+
+                result.ExitCode.Should().Be(exitCode, because, becauseArgs);
+
+                if (string.IsNullOrEmpty(stdOutPattern))
+                    result.StdOut.Should().BeEmpty(because, becauseArgs);
+                else
+                    result.StdOut.Should().MatchEquivalentOf(stdOutPattern, opt => opt.IgnoringNewlineStyle(), because, becauseArgs);
+
+                if (string.IsNullOrEmpty(stdErrPattern))
+                    result.StdErr.Should().BeEmpty(because, becauseArgs);
+                else
+                    result.StdErr.Should().MatchEquivalentOf(stdErrPattern, opt => opt.IgnoringNewlineStyle(), because, becauseArgs);
+
                 failures = scope.Discard();
 
                 return failures.Length == 0;
@@ -92,42 +99,6 @@ public class NugraphAssertions(NugraphResult instance, AssertionChain chain) :
 
                 using var scope = new AssertionScope();
                 diagram.Should().BeEquivalentTo(expectedDiagram, opt => opt.IgnoringNewlineStyle(), because, becauseArgs);
-                failures = scope.Discard();
-
-                return failures.Length == 0;
-            })
-            .FailWith(string.Join(Environment.NewLine, failures.Select(e => e.Replace("{", "{{").Replace("}", "}}"))));
-
-        return new AndConstraint<NugraphAssertions>(this);
-    }
-
-    /// <summary>
-    /// Asserts that the nugraph result is a failure.
-    /// <list type="bullet">
-    /// <item>Ensures that exit code is <paramref name="failCode"/>.</item>
-    /// <item>Ensures that stderr matches the provided pattern.</item>
-    /// </list>
-    /// </summary>
-    /// <param name="failCode">The expected exit code.</param>
-    /// <param name="stdErrPattern">The expected pattern that must be written on stderr.</param>
-    /// <param name="because">
-    /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
-    /// </param>
-    /// <param name="becauseArgs">
-    /// Zero or more objects to format using the placeholders in <paramref name="because" />.
-    /// </param>
-    [CustomAssertion]
-    public AndConstraint<NugraphAssertions> Fail(int failCode, string stdErrPattern, string because = "", params object[] becauseArgs)
-    {
-        string[] failures = [];
-
-        _chain
-            .Given(() => Subject)
-            .ForCondition(result =>
-            {
-                using var scope = new AssertionScope();
-                result.ExitCode.Should().Be(failCode, because, becauseArgs);
-                result.StdErr.Should().MatchEquivalentOf(stdErrPattern, opt => opt.IgnoringNewlineStyle(), because, becauseArgs);
                 failures = scope.Discard();
 
                 return failures.Length == 0;